@@ -4,7 +4,7 @@ version: 2.0.0-nullsafety.3
 homepage: https://github.com/fysoul17/google_maps_place_picker
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: '>=2.18.0 <3.0.0'
 
 dependencies:
   flutter:
